parser: move type string logic into separate module
